requirements.txt
async-timeout==3.0.1
beautifulsoup4==4.8.2
discord-py-slash-command==2.0.4
discord.py==1.7.3
pandas==1.2.3
python-dotenv==0.18.0
requests==2.25.1